Version 2.36
Bugfix by reverting a file to 2.32 version. The bugs caused by 2.33 were due to `self._player_slots = {}` in `_missionbriefinggui` which is avoidable but the "reduction" /hook did not provide much value. With `_menulobbyrenderer`, the issue was that `self:_entered_menu()` gets called at the end preventing an easy reduction via hook. Bad Overkill!