Fix CTD when going into Soundtrack menus - Fixes #52 (#66)
* Fix CTD when going into Soundtrack menus

Fixes CTS when going into soundtrackmenus by cancelling the initialization if the network manager has no session. This behaviour is also found in the original code, but will not avoid the added code from biglobby from being executed.

For some reason, a "Waiting for Server" will appear in the lower right corner when entereing the menu though. But I expect that to be caused by something different.

* Extra fix concerning the sound menu

I was still crashing in the sound menu when using WolfHUD, so I kept looking for the cause. Turns out you chained the wrong originial function into a function call.